requirements.txt
flask
gspread
oauth2client
pytesseract
Pillow
